# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API",
        "description": "API Documentation",
        "contact": {
            "email": "ekoabdulgoffar129@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:49/bpm-backend-laravel/api/v1",
            "description": "API Documentation"
        }
    ],
    "paths": {
...
        "/api/v1/template_survei": {
            "get": {
                "tags": [
                    "TemplateSurvei"
                ],
                "summary": "Get list of template survei",
                "operationId": "8e16ae798c519ad836a96c5ef19a7ab8",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "tsu_id": {
                                                "type": "integer"
                                            },
                                            "tsu_nama": {
                                                "type": "string"
                                            },
                                            "tsu_status": {
                                                "type": "string"
                                            },
                                            "tsu_created_by": {
                                                "type": "string"
                                            },
                                            "tsu_created_date": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "tsu_modif_by": {
                                                "type": "string"
                                            },
                                            "tsu_modif_date": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "ksr_id": {
                                                "type": "integer"
                                            },
                                            "skp_id": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "TemplateSurvei"
                ],
                "summary": "Create new template survei",
                "operationId": "54ee405557022ff5832d6553a61212e2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "tsu_nama": {
                                        "type": "string"
                                    },
                                    "tsu_status": {
                                        "type": "string"
                                    },
                                    "tsu_created_by": {
                                        "type": "string"
                                    },
                                    "tsu_created_date": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "tsu_modif_by": {
                                        "type": "string"
                                    },
                                    "tsu_modif_date": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "ksr_id": {
                                        "type": "integer"
                                    },
                                    "skp_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Template Survei created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/template_survei/{id}": {
            "get": {
                "tags": [
                    "TemplateSurvei"
                ],
                "summary": "Get template survei by ID",
                "operationId": "0398881c3faf03bb2267cc50a2eb62b8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID template survei",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "tsu_id": {
                                            "type": "integer"
                                        },
                                        "tsu_nama": {
                                            "type": "string"
                                        },
                                        "tsu_status": {
                                            "type": "string"
                                        },
                                        "tsu_created_by": {
                                            "type": "string"
                                        },
                                        "tsu_created_date": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "tsu_modif_by": {
                                            "type": "string"
                                        },
                                        "tsu_modif_date": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "ksr_id": {
                                            "type": "integer"
                                        },
                                        "skp_id": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Template Survei not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "TemplateSurvei"
                ],
                "summary": "Update template survei",
                "operationId": "4cb8c77b342ced979cb0bc72b3feda5c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID template survei",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "tsu_nama": {
                                        "type": "string"
                                    },
                                    "tsu_status": {
                                        "type": "string"
                                    },
                                    "tsu_created_by": {
                                        "type": "string"
                                    },
                                    "tsu_created_date": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "tsu_modif_by": {
                                        "type": "string"
                                    },
                                    "tsu_modif_date": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "ksr_id": {
                                        "type": "integer"
                                    },
                                    "skp_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Template Survei updated successfully"
                    },
                    "404": {
                        "description": "Template Survei not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "TemplateSurvei"
                ],
                "summary": "Delete template survei",
                "operationId": "85793d85ca1d9af211f816f469eeb02e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID template survei",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Template Survei deleted successfully"
                    },
                    "404": {
                        "description": "Template Survei not found"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token",
                "scheme": "bearer"
            }
        }
    },
...
}